+++ pkg/services/ngalert/notifier/legacy_storage/compat_test.go
@@ -10,8 +10,12 @@ import (
 	"github.com/grafana/alerting/notify/notifytest"
 	"github.com/grafana/alerting/receivers/schema"
 	"github.com/grafana/alerting/receivers/teams"
+	"github.com/prometheus/common/model"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	"github.com/grafana/grafana/pkg/services/ngalert/api/tooling/definitions"
+	"github.com/grafana/grafana/pkg/services/ngalert/models"
 )
 
 func TestPostableMimirReceiverToIntegrations(t *testing.T) {
@@ -56,4 +60,31 @@ func TestPostableMimirReceiverToIntegrations(t *testing.T) {
 		require.NoError(t, err)
 		require.Empty(t, integrations)
 	})
+}
+
+func TestManagedRouteToRoute(t *testing.T) {
+	gw := model.Duration(10)
+	gi := model.Duration(20)
+	ri := model.Duration(30)
+
+	mr := &ManagedRoute{
+		Name:           "test",
+		Receiver:       "receiver",
+		GroupBy:        []string{"alertname"},
+		GroupWait:      &gw,
+		GroupInterval:  &gi,
+		RepeatInterval: &ri,
+		Routes:         []*definition.Route{{Receiver: "child"}},
+		Provenance:     models.Provenance("test"),
+	}
+
+	route := ManagedRouteToRoute(mr)
+
+	assert.Equal(t, "receiver", route.Receiver)
+	assert.Equal(t, []string{"alertname"}, route.GroupByStr)
+	assert.Equal(t, &gw, route.GroupWait)
+	assert.Equal(t, &gi, route.GroupInterval)
+	assert.Equal(t, &ri, route.RepeatInterval)
+	assert.Len(t, route.Routes, 1)
+	assert.EqualValues(t, definitions.Provenance("test"), route.Provenance)
 }
